backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.46
pydantic==1.10.7
psycopg2-binary>=2.9.10
alembic==1.10.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.59.7